switch from float to 0.1 deg C to avoid float sideeffects delay activation of heater for 30sec to allow pumps and other things come up first

+++ Temperature.h
@@ -23,9 +23,9 @@ public:
     void handleTick();
     DeviceId getId();
     DeviceType getType();
-    float getMinimum();
-    float getMaximum();
-    float getSensorTemperature(byte[]);
+    int16_t getMinimum();
+    int16_t getMaximum();
+    int16_t getSensorTemperature(byte[]);
 
 protected:
